pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legalqa"
version = "0.1.0"
description = "Legal Question Answering System"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    # Core application dependencies
    "pandas",
    "pyarrow",
    "python-dotenv",
    "faiss-cpu",
    "langchain",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "rank-llm",
    "fastapi",
    "uvicorn[standard]", # uvicorn is needed to run fastapi
    "psycopg2-binary",
    "pgvector",
    "scikit-learn",
    "numpy",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy",
]
notebook = [
    "matplotlib",
    "seaborn",
    "jupyter",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.hatch.build.targets.wheel]
packages = ["src"] 